+++ CHANGELOG.md
@@ -13,6 +13,8 @@ about writing changes to this log.
 - [TVIST1-725](https://jira.itkdev.dk/browse/TVIST1-725):
   - Allow deletion of created hearing post requests and responses
   - Send hearing post response receipt upon approval rather than creation.
+- [TVIST1-729](https://jira.itkdev.dk/browse/TVIST1-729):
+  Allowed hearing to start with just a party or counterparty.
 
 ## [1.0.3] 2022-12-19